feat: rename aleph.sdk.chains.sol to aleph.sdk.chains.solana
